model abstraction for GET
